% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsparrow_model.R
\name{print.rsparrow_data}
\alias{print.rsparrow_data}
\title{Print an rsparrow_data Object}
\usage{
\method{print}{rsparrow_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rsparrow_data"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an rsparrow_data Object
}
