% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsparrow-package.R
\docType{package}
\name{rsparrow-package}
\alias{rsparrow-package}
\alias{rsparrow}
\title{rsparrow: SPARROW Water Quality Modeling in R}
\description{
Implements the USGS SPARROW (SPAtially Referenced Regressions On Watershed
attributes) model for estimating contaminant sources, watershed delivery,
and in-stream transport in river networks.
}
\section{Main Functions}{

\itemize{
\item \code{\link{rsparrow_model}}: Estimate a SPARROW model
\item \code{\link{predict.rsparrow}}: Predict loads and yields
\item \code{\link{summary.rsparrow}}: Summarize estimation results
\item \code{\link{rsparrow_hydseq}}: Compute hydrological sequencing
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://code.usgs.gov/water/stats/RSPARROW}
}

}
\author{
\strong{Maintainer}: Kyle Hurley \email{khurley@usgs.gov}

Authors:
\itemize{
  \item Lillian Gorman Sanisaca \email{lgormansanisaca@usgs.gov}
  \item Richard Alexander \email{ralex@usgs.gov}
}

}
\keyword{internal}
